% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_two_point_data}
\alias{example_two_point_data}
\title{Example Data: Two Point Data}
\format{
\subsection{\code{example_two_point_data}}{

A dataframe with 10 rows and 30 columns:
\describe{
\item{Animal}{The animal name/unique identifier}
\item{Time}{The time of the measurement, in days.}
\item{SD.DR}{The name of the replicate column.}
}
}
}
\source{
This data was extracted from Romero & Rich 2007 (Comp Biochem. Physiol. Part A Mol. Integr. Physiol. 147, 562-568. https://doi.org/10.1016/j.cbpa.2007.02.004)
}
\usage{
example_two_point_data
}
\description{
An example of data that one would perform profile repeatability on.
Consists of 9 individual animals, with corticosterone data taken at 2 timepoints (n_trials = 2), baseline (time = 3) and stress-induced (time = 30).
Then, there are 28 replicate columns.
}
\keyword{datasets}
