% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\title{Calculates the variances of the groups}
\usage{
get_vars(individual_array, n_replicates)
}
\arguments{
\item{individual_array}{The array of data for an individual}

\item{n_replicates}{The number of replicate groups}
}
\value{
A list, where the elements are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{     1. variances: A vector of the variances of the sample
     2. total_sum: The sum of all the measurements in the sample
     3. ssq: The sum of all the squares of the measurements in the sample
     4. num_measurements: The total number of measurements in the sample that are not null
}\if{html}{\out{</div>}}
}
\description{
For the given individual i, for a given time, computes the variance in values
over replicates.
}
\details{
Returns these variances, sum of all values (for all times and replicates),
sum of all these values squared, and the number of values.

Change: instead of passing in i, individuals -> let's do individual
Then whatever calls this function should do individuals at i
}
