% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_individual_df.R
\name{score_individual_df}
\alias{score_individual_df}
\title{Score an Individual Data Frame}
\usage{
score_individual_df(
  individual_df,
  n_trials,
  n_replicates,
  max_variance,
  variance_set
)
}
\arguments{
\item{individual_df}{A data frame containing individual data.}

\item{n_trials}{The total number of trials in the data frame.}

\item{n_replicates}{The total number of replicates in each trial.}

\item{max_variance}{The maximum allowed variance value.}

\item{variance_set}{A vector of variance values.}
}
\value{
A score calculated for the individual data frame.
}
\description{
This function calculates a score for an individual data frame based on various factors,
including the number of crossovers, maximum variance, and a set of variances.
}
\details{
The score is computed as follows:
\itemize{
\item It factors in the number of crossovers using a scaling factor.
\item It considers the maximum variance value in the variance set.
\item It adds a component based on the average of variance values.
\item It includes a scaled component of the number of crossovers.
}
}
\seealso{
\code{\link{calculate_crossovers}} for information on crossovers calculation.
}
