% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_helpers.R
\name{sigmoid}
\alias{sigmoid}
\title{Calculates the sigmoid function of the input}
\usage{
sigmoid(float)
}
\arguments{
\item{float}{A float number}
}
\value{
A float number which is the result of the sigmoid function
}
\description{
Calculates the sigmoid function of the input
}
\examples{
sigmoid(0)
sigmoid(2)

}
