% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sparrow_repeatability_three_point}
\alias{sparrow_repeatability_three_point}
\title{Example Data: Sparrow Repeatability (3 Point Data)}
\format{
\subsection{\code{sparrow_repeatability_three_point}}{

A dataframe with 36 rows and 12 columns:
\describe{
\item{Animal}{The animal name/unique identifier}
\item{TIME}{The time of the measurement, in days}
\item{LD.500}{The name of the replicate column}
}
}
}
\source{
An experiment run at Tufts University by Ursula K. Beattie
}
\usage{
sparrow_repeatability_three_point
}
\description{
An example of data that one would perform profile repeatability on.
Consists of 12 individual animals, with data taken at 3 times (n_trials = 3).
Then, there are 10 replicate columns. This example also shows what happens
when there are null data records for some individuals.
}
\keyword{datasets}
