% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_data_four_point}
\alias{synthetic_data_four_point}
\title{Example Data: Synthetic 4-Point Data}
\format{
\subsection{\code{synthetic_data_four_point}}{

A dataframe with 44 rows and 6 columns:
\describe{
\item{Animal}{The animal name/unique identifier}
\item{TIME}{The time of the measurement (unit not important)}
\item{A}{The (unimportant) name of a replicate column.}
\item{B}{The (unimportant) name of a replicate column.}
\item{C}{The (unimportant) name of a replicate column.}
\item{D}{The (unimportant) name of a replicate column.}
}
}
}
\source{
Data created for testing purposes by
}
\usage{
synthetic_data_four_point
}
\description{
An example of data that one would perform profile repeatability on.
The data is synthetic data created for testing purposes.
Consists of 11 individual animals, with data taken at 4 times (n_trials = 4).
Then, there are four replicate columns. Replicate column names refer to
sample tests performed on the animal.
}
\keyword{datasets}
