% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ordering.R
\name{do_ordering}
\alias{do_ordering}
\title{Performs the ordering of input data by scoring each individual data frame.}
\usage{
do_ordering(n_trials, id_list, df_list, n_replicates)
}
\arguments{
\item{n_trials}{The number of trials per individual (number of rows per individual dataframe)}

\item{id_list}{The list of unique individual names}

\item{df_list}{the list of dataframe per unique individual}

\item{n_replicates}{The number of replicates for each individual}
}
\description{
Performs the ordering of input data by scoring each individual data frame.
}
