% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_two_point_data}
\alias{example_two_point_data}
\title{Example Data: Two Point Data}
\format{
\subsection{\code{example_two_point_data}}{

A dataframe with 10 rows and 30 columns:
\describe{
\item{Animal}{The animal name/unique identifier}
\item{Time}{The time of the measurement, in days.}
\item{SD.DR}{The name of the replicate column.}
}
}
}
\source{
An experiment run at Tufts University by Ursula K. Beattie
}
\usage{
example_two_point_data
}
\description{
An example of data that one would perform profile repeatability on.
Consists of 9 individual animals, with data taken at 2 times (n_trials = 2).
Then, there are 28 replicate columns.
}
\keyword{datasets}
