% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profrep.R
\name{profrep}
\alias{profrep}
\title{Perform profile repeatability on a data-frame}
\usage{
profrep(df, n_trials)
}
\arguments{
\item{df}{The data frame Column 1 is the animal, column 2 is the time, the rest are the data}

\item{n_trials}{The number of rows an individual animal will have.}
}
\value{
Does not return any object, but prints out the summary of the
profile repeatability measure for the data. TODO: Change this to return
a dataframe.
}
\description{
Perform profile repeatability on a data-frame
}
