% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profrep.R
\name{profrep}
\alias{profrep}
\title{Perform profile repeatability}
\usage{
profrep(items, n_rows)
}
\arguments{
\item{items}{The data file}

\item{n_rows}{The number of rows}
}
\value{
Does not return any object, but prints out the summary of the
profile repeatability measure for the data. TODO: Change this to return
a dataframe.
}
\description{
\code{profrep()} performs the profile repeatability metric on the provided item
}
