% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dfs.R
\name{score_dfs}
\alias{score_dfs}
\title{Compute the profile repeatability score for each groups of data}
\usage{
score_dfs(id_list, df_list, n_replicates, n_trials)
}
\arguments{
\item{id_list}{The list of the names of the individuals}

\item{df_list}{A list of data frames, each of which correspond to one of the names in the individual list}

\item{n_replicates}{The number of replicate columns (number of columns in a df in df_list)}

\item{n_trials}{The number of trials per individual (number of rows in a df in df_list)}
}
\value{
A dataframe of two columns, where the first column is the individual name and the second is the profile repeatability score for the individual's data frame
}
\description{
Splits the data into the data frame for a particular individual from
the id_list, then calculates metrics to compute the profile repeatability
score. Returns a data frame with the individuals name and the score.
}
\examples{
id_list <- c('A', 'B', 'C')
df_list <- c(
  matrix(c(1, 60, 1, 2, 3, 4, 5, 1, 90, 9, NA, 4, NA, 2, 1, 120, 3, 6, NA, NA, 9), nrow = 3, byrow=TRUE),
  matrix(c(1, 60, 5, 6, 7, 8, 9, 1, 90, 9, NA, 4, NA, 2, 1, 120, 3, 6, NA, NA, 9), nrow = 3, byrow=TRUE),
  matrix(c(1, 60, 11, 12, 13, 14, 15, 1, 90, 9, NA, 4, NA, 2, 1, 120, 3, 6, NA, NA, 9), nrow = 3, byrow=TRUE)
  )
ret_df <- score_dfs(id_list, df_list, n_replicates=5, n_trials=3)

}
