% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sparrow_repeatability_three_point}
\alias{sparrow_repeatability_three_point}
\title{Example Data: Sparrow Repeatability (3 Point Data)}
\format{
\subsection{\code{sparrow_repeatability_three_point}}{

A dataframe with 36 rows and 12 columns:
\describe{
\item{Animal}{The animal name/unique identifier}
\item{TIME}{The time of the measurement, in days}
\item{LD.500}{The name of the replicate column}
}
}
}
\source{
This data was extracted from Rich & Romero 2001 (J. Comp. Physiol. Part B Biochem. Syst. Environ. Physiol. 171, 543-647. https://doi.org/10.1007/s003600100204)
}
\usage{
sparrow_repeatability_three_point
}
\description{
An example of data that one would perform profile repeatability on.
Consists of 12 individual animals, with corticosterone data taken at 3 times (n_trials = 3), baseline (time = 0) and two stress-induced (time = 15 and 30).
Then, there are 10 replicate columns. This example also shows what happens
when there are null data records for some individuals.
}
\keyword{datasets}
